requirements.txt
apache-airflow[celery,crypto,emr,hive,hdfs,ldap,mysql,postgres,redis,slack,s3,kubernetes]==1.10.2
# git+git://github.com/apache/airflow.git@de75b7a2bd7f5bef6a1d09942e0b43c17a3fbb95#egg=master # https://stackoverflow.com/questions/20101834/pip-install-from-git-repo-branch/20101940#20101940
celery[redis]>=4.1.1,<4.2.0
cython==0.29
alembic==0.8.10
certifi==2018.4.16
chardet==3.0.4
click==6.7
croniter==0.3.24
dill==0.2.8.2
docutils==0.14
Flask==0.12.4
flask-admin==1.5.2
Flask-Cache==0.13.1
Flask-Login<0.5,>=0.3
flask-swagger==0.2.13
Flask-WTF==0.14.2
funcsigs==1.0.0
future==0.16.0
gitdb2==2.0.4
GitPython==2.1.11
gunicorn==19.5.0
idna==2.7
itsdangerous==0.24
Jinja2==2.8.1
lockfile==0.12.2
lxml>=4.0.0
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.15.0
ordereddict==1.1
pandas==0.23.3
psutil==4.4.2
Pygments==2.2.0
python-daemon==2.1.2
python-dateutil==2.7.3
python-editor==1.0.3
python-nvd3==0.15.0
python-slugify==1.2.5
pytz==2018.5
PyYAML==3.13
requests<3,>=2.20.0
setproctitle==1.1.10
six==1.11.0
smmap2==2.0.4
SQLAlchemy==1.1.15
tabulate==0.7.7
thrift==0.9.3
Unidecode==1.0.22
urllib3==1.23
Werkzeug==0.14.1
WTForms==2.2.1
zope.deprecation==4.3.0
pyOpenSSL==18.0.0
ndg-httpsclient==0.5.1
pyasn1==0.4.4
newrelic==4.6.0.106
newrelic-api